refactor store, fixes #35
